+++ libs/react-components/src/community/components/table-of-contents/table-of-contents.stories.tsx
@@ -32,6 +32,7 @@ const meta: Meta<TableOfContentsProps> = {
 
 export default meta;
 
+faker.seed(123);
 const steps = [...Array(10).keys()].map(() => faker.commerce.productName());
 const longSteps = [...steps, ...steps];