+++ libs/react-components/src/tedi/components/popover/popover-trigger.tsx
@@ -1,0 +1,5 @@
+import { OverlayTrigger, OverlayTriggerProps } from '../overlay/overlay-trigger';
+
+export function PopoverTrigger(props: OverlayTriggerProps) {
+  return <OverlayTrigger {...props} />;
+}