+++ libs/react-components/src/tedi/helpers/index.ts
@@ -4,3 +4,4 @@ export * from './hooks/use-breakpoint-props';
 export * from './hooks/use-print';
 export * from './hooks/use-element-size';
 export * from './hooks/use-scroll';
+export * from './hooks/use-is-touch-device';